Add static json definition for Power Query standard library (#4)
This PR creates a new local package (powerquery-library) which will be responsible for providing metadata for all M functions, types, and constants. In this short term this is loaded from a static .json file. Longer term this will be dynamically retrieved from a local engine instance.

The current code takes a simplistic approach and returns all module exports from the file. This could be further refined to return a limited set of modules (i.e. Power BI specific, available to the SDK, etc).

The library file currently contains definitions for most of the modules that are available in Power Query for Power BI (but is missing a number of data source modules).