Move intellisense logic into language-services package (#25)
This change refactors the intellisense code that was previously implemented in the server project into a new (standalone) package - powerquery-language-services. The interfaces involved (Analysis, AnalysisOptions, SymbolProvider, WorkspaceCache) are not final, but are good enough to maintain parity with the current vscode extension.

The only new behavior is the inclusion of keywords in the completion item list (#23). I decided to integrate that as part of this change as a way of proving out the use of multiple SymbolProviders.